Crate automaton_test

Project intended for incremental building of a state machine. The intention is to make every language item presented to the developer as explicit as possible. While still allowing some degree of dynamic flow. Only using safe code of-course!

Modules

function
Contains the core functionality items for our system.

marker
Primitive traits which can be used as constraints by the core components.

service
Types which attribute functionality to state machines.

state
Types which encode the states to be used by a state machine.

stm
Traits enforcing state machine behaviour.

transaction
Types used to convey transition related information.

Structs

Machine
The state machine.