Trait automaton_test::function::error::SnapshottedErrorExt

pub trait SnapshottedErrorExt<T> {
    fn context<M>(
        self, 
        context: ErrorKind, 
        machine: &M
    ) -> Result<T, MachineError>
    where
        M: StateContainer + Clone + Debug + Sync + Send + 'static;
}

Trait facilitating error creation with a snapshot of the state machine attached.

Required Methods

fn context<M>(self, context: ErrorKind, machine: &M) -> Result<T, MachineError> where
M: StateContainer + Clone + Debug + Sync + Send + 'static,

Converts the Error type of a Result object into a MachineError.