Trait automaton_test::function::State

pub trait State {
    type Transaction;
}

Trait generalizing over any state that's present in the state machine.

Associated Types

type Transaction

Type of structure which must be provided when transitioning into the state represented by the enclosing type.