Trait automaton_test::stm::PullupFrom

pub trait PullupFrom<T, TTC> where
    TTC: TransactionContainer + 'static,
    T: StateContainer + ServiceCompliance<StackStorage<TTC>> + Sized + 'static,
    Self: Sized + StateContainer + 'static,
    Self::State: State,
    <Self::State as State>::Transaction: Transaction + 'static, {
    fn pullup_from(_: T) -> Result<Self, MachineError>;
}

Types, state machines residing in a certain state, which transform one-sided into a previous Type. The Transaction object of the next state is loaded and restored.

PullupFrom is designed to be used together with PushdownFrom because one part of it's functionality is to restore the next state's Transaction from a stack. Generally each [PushDown] must be followed with a matching Pullup operation to correctly push onto and pop from the stackstorage.

A state machine is said to pullup from B into A when the current state is B and the following transition is valid [A <- B].

Required Methods

fn pullup_from(_: T) -> Result<Self, MachineError>

Transition from the provided state into the implementing state.

Errors

There is a check at runtime which prevents a Pullup transition if it doesn't match the correct PushDown transition in a First In, Last Out (FILO) manner. Note: This part CANNOT be statically verified as far as I know?