Trait automaton_test::stm::PushdownFrom

pub trait PushdownFrom<T, TTC> where
    TTC: TransactionContainer + 'static,
    T: StateContainer + Sized + 'static,
    Self: StateContainer + ServiceCompliance<StackStorage<TTC>> + Sized + 'static,
    Self::State: State,
    <Self::State as State>::Transaction: Transaction + Copy + 'static, {
    fn pushdown_from(_: T, _: <Self::State as State>::Transaction) -> Self;
}

Types, state machines residing in a certain state, which transform one-sided into a next Type. The Transaction object of the previous state is stored for re-use.

PushdownFrom is designed to be used together with PullupFrom because one part of it's functionality is to store the previous state's Transaction onto a stack. Generally each [PushDown] must be followed with a matching Pullup operation to correctly push onto and pop from the stackstorage.

A state machine is said to pushdown from A into B when the current state is A, a Transaction object for state B is provided and the following transition is valid [A -> B].

Required Methods

fn pushdown_from(_: T, _: <Self::State as State>::Transaction) -> Self

Transition from the provided state into the implementing state.