Struct automaton_test::Machine

pub struct Machine<X> where
    X: TopLevelMarker + State,  {
    pub state: PhantomData<X>,
    pub transaction: X::Transaction,
    pub storage: StackStorage<TransactionItem>,
}

The state machine.

The developer is encouraged to design this structure in any desired way by storing services into it's members. Each state machine MUST have a state and transaction field AT MINIMUM.

Fields
state: PhantomData<X>

Field to encode the current state of the machine.

This field is present to utilize the type system to statically verify legal transitions of the machine. This field has no (/zero) size at runtime.

transaction: X::Transaction

Field to store the provided Transaction object as rquired by the current state.

storage: StackStorage<TransactionItem>

Stack storage service to allow PushDown and Pullup behaviour to be implemented.